Test for terminal prior to asking to upgrade
Apt does not like when a terminal is missing. We'll skip all install
parts unless a terminal is present. Additionally, we'll download the
full updates as well, without even pausing to ask if it is okay.

If you wanted to be asked, you would have set up a terminal first.